# web/src/lib/generated/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AssetFlow API",
    "version": "0.1.0"
  },
  "paths": {
    "/api/auth/register": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Register",
        "operationId": "register_api_auth_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Login",
        "operationId": "login_api_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/assets": {
      "get": {
        "tags": [
          "assets"
        ],
        "summary": "Get Positions",
        "operationId": "get_positions_api_assets_get",
        "parameters": [
          {
            "name": "assetflow_access",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Assetflow Access"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AssetPosition"
                  },
                  "title": "Response Get Positions Api Assets Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "assets"
        ],
        "summary": "Add Asset",
        "operationId": "add_asset_api_assets_post",
        "parameters": [
          {
            "name": "assetflow_access",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Assetflow Access"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssetCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssetPosition"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/assets/{asset_id}": {
      "patch": {
        "tags": [
          "assets"
        ],
        "summary": "Patch Asset",
        "operationId": "patch_asset_api_assets__asset_id__patch",
        "parameters": [
          {
            "name": "asset_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Asset Id"
            }
          },
          {
            "name": "assetflow_access",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Assetflow Access"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssetUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssetPosition"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "assets"
        ],
        "summary": "Delete Asset",
        "operationId": "delete_asset_api_assets__asset_id__delete",
        "parameters": [
          {
            "name": "asset_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Asset Id"
            }
          },
          {
            "name": "assetflow_access",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Assetflow Access"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/assets/{asset_id}/price": {
      "post": {
        "tags": [
          "assets"
        ],
        "summary": "Patch Asset Price",
        "operationId": "patch_asset_price_api_assets__asset_id__price_post",
        "parameters": [
          {
            "name": "asset_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Asset Id"
            }
          },
          {
            "name": "assetflow_access",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Assetflow Access"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssetPriceUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssetPosition"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/assets/{asset_id}/price/sync": {
      "post": {
        "tags": [
          "assets"
        ],
        "summary": "Sync Asset Price",
        "operationId": "sync_asset_price_api_assets__asset_id__price_sync_post",
        "parameters": [
          {
            "name": "asset_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Asset Id"
            }
          },
          {
            "name": "assetflow_access",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Assetflow Access"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssetPosition"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/assets/{asset_id}/chart": {
      "get": {
        "tags": [
          "assets"
        ],
        "summary": "Asset Chart",
        "operationId": "asset_chart_api_assets__asset_id__chart_get",
        "parameters": [
          {
            "name": "asset_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Asset Id"
            }
          },
          {
            "name": "range_days",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 365,
              "minimum": 7,
              "default": 30,
              "title": "Range Days"
            }
          },
          {
            "name": "assetflow_access",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Assetflow Access"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssetChartResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/assets/providers/search": {
      "get": {
        "tags": [
          "assets"
        ],
        "summary": "Asset Provider Search",
        "operationId": "asset_provider_search_api_assets_providers_search_get",
        "parameters": [
          {
            "name": "asset_type",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/AssetType"
            }
          },
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 2,
              "title": "Q"
            }
          },
          {
            "name": "assetflow_access",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Assetflow Access"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AssetInstrumentOption"
                  },
                  "title": "Response Asset Provider Search Api Assets Providers Search Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AssetChartResponse": {
        "properties": {
          "asset_id": {
            "type": "string",
            "title": "Asset Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "range_days": {
            "type": "integer",
            "title": "Range Days"
          },
          "points": {
            "items": {
              "$ref": "#/components/schemas/AssetChartPoint"
            },
            "type": "array",
            "title": "Points"
          }
        },
        "type": "object",
        "required": [
          "asset_id",
          "name",
          "range_days",
          "points"
        ],
        "title": "AssetChartResponse"
      },
      "AssetCreate": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "type": {
            "$ref": "#/components/schemas/AssetType"
          },
          "currency": {
            "type": "string",
            "title": "Currency"
          },
          "symbol": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Symbol"
          },
          "quantity": {
            "type": "number",
            "title": "Quantity",
            "default": 0
          },
          "average_buy_price": {
            "type": "number",
            "title": "Average Buy Price",
            "default": 0
          },
          "average_buy_price_in_base": {
            "type": "number",
            "title": "Average Buy Price In Base",
            "default": 0
          },
          "current_price": {
            "type": "number",
            "title": "Current Price",
            "default": 0
          },
          "current_price_in_base": {
            "type": "number",
            "title": "Current Price In Base",
            "default": 0
          },
          "invested_amount_in_base": {
            "type": "number",
            "title": "Invested Amount In Base",
            "default": 0
          },
          "risk_label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Risk Label"
          },
          "liquidity_label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Liquidity Label"
          },
          "tracking_enabled": {
            "type": "boolean",
            "title": "Tracking Enabled",
            "default": false
          },
          "tracking_provider": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tracking Provider"
          },
          "tracking_external_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tracking External Id"
          },
          "tracking_symbol": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tracking Symbol"
          },
          "rental_enabled": {
            "type": "boolean",
            "title": "Rental Enabled",
            "default": false
          },
          "rental_income_monthly": {
            "type": "number",
            "title": "Rental Income Monthly",
            "default": 0
          },
          "rental_payment_frequency": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rental Payment Frequency"
          },
          "rental_payment_day": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rental Payment Day"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          }
        },
        "type": "object",
        "required": [
          "name",
          "type",
          "currency"
        ],
        "title": "AssetCreate"
      },
...
      "AssetPosition": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "symbol": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Symbol"
          },
          "type": {
            "$ref": "#/components/schemas/AssetType"
          },
          "currency": {
            "type": "string",
            "title": "Currency"
          },
          "quantity": {
            "type": "number",
            "title": "Quantity"
          },
          "current_price": {
            "type": "number",
            "title": "Current Price"
          },
          "current_price_in_base": {
            "type": "number",
            "title": "Current Price In Base"
          },
          "current_value_in_base": {
            "type": "number",
            "title": "Current Value In Base"
          },
          "invested_amount_in_base": {
            "type": "number",
            "title": "Invested Amount In Base"
          },
          "unrealized_pnl": {
            "type": "number",
            "title": "Unrealized Pnl"
          },
          "allocation_pct": {
            "type": "number",
            "title": "Allocation Pct"
          },
          "tracking_enabled": {
            "type": "boolean",
            "title": "Tracking Enabled",
            "default": false
          },
          "tracking_provider": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tracking Provider"
          },
          "tracking_external_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tracking External Id"
          },
          "tracking_symbol": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tracking Symbol"
          },
          "linked_account_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Linked Account Id"
          },
          "linked_account_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Linked Account Name"
          },
          "valuation_source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Valuation Source"
          },
          "rental_enabled": {
            "type": "boolean",
            "title": "Rental Enabled",
            "default": false
          },
          "rental_income_monthly": {
            "type": "number",
            "title": "Rental Income Monthly",
            "default": 0
          },
          "rental_payment_frequency": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rental Payment Frequency"
          },
          "rental_payment_day": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rental Payment Day"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "symbol",
          "type",
          "currency",
          "quantity",
          "current_price",
          "current_price_in_base",
          "current_value_in_base",
          "invested_amount_in_base",
          "unrealized_pnl",
          "allocation_pct"
        ],
        "title": "AssetPosition"
      },
      "AssetPriceUpdate": {
        "properties": {
          "priced_at": {
            "type": "string",
            "format": "date",
            "title": "Priced At"
          },
          "price": {
            "type": "number",
            "title": "Price"
          },
          "price_in_base": {
            "type": "number",
            "title": "Price In Base"
          },
          "source": {
            "type": "string",
            "title": "Source",
            "default": "manual"
          }
        },
        "type": "object",
        "required": [
          "priced_at",
          "price",
          "price_in_base"
        ],
        "title": "AssetPriceUpdate"
      },
      "AssetType": {
        "type": "string",
        "enum": [
          "cash",
          "deposit",
          "stock",
          "etf",
          "bond",
          "crypto",
          "metal",
          "real_estate",
          "custom"
        ],
        "title": "AssetType"
      },
...
      "AuthResponse": {
        "properties": {
          "user": {
            "$ref": "#/components/schemas/UserSummary"
          }
        },
        "type": "object",
        "required": [
          "user"
        ],
        "title": "AuthResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LoginRequest": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "password": {
            "type": "string",
            "minLength": 8,
            "title": "Password"
          }
        },
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "title": "LoginRequest"
      },
...
      "MessageResponse": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "message"
        ],
        "title": "MessageResponse"
      },
...
      "RegisterRequest": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "full_name": {
            "type": "string",
            "maxLength": 255,
            "minLength": 2,
            "title": "Full Name"
          },
          "password": {
            "type": "string",
            "minLength": 8,
            "title": "Password"
          },
          "base_currency": {
            "type": "string",
            "title": "Base Currency",
            "default": "USD"
          },
          "timezone": {
            "type": "string",
            "title": "Timezone",
            "default": "UTC"
          }
        },
        "type": "object",
        "required": [
          "email",
          "full_name",
          "password"
        ],
        "title": "RegisterRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}